% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{new_surveys_wide}
\alias{new_surveys_wide}
\title{New surveys data for the next congress elections in a long
format}
\format{
A tibble with 10 variables:
\itemize{
\item \code{id_survey}: survey's id constructed from the
polling firm and the fieldwork dates.
\item \code{id_elec}: election's id constructed from the
election code and date.
\item \code{polling_firm}: polling firm in charge of conducting
the poll.
\item \code{media}: media outlet that published the survey.
\item \code{fieldwork_start}: fieldwork start date.
\item \code{fieldwork_end}: fieldwork end date.
\item \code{n_field_days}: length of fieldwork days.
\item \code{sample_size}: sample size of the survey.
\item \code{abbrev_candidacies}: party abbreviation.
\item \code{estimated_voting}: estimated vote.
}
}
\source{
Data extracted and updated from press reports published by
polling firms, readily available on open-access pages such as
Wikipedia
}
\usage{
data(new_surveys_wide)
}
\description{
A dataset containing surveys about the forthcoming
congress elections in Spain, daily updated via Github action
}
\examples{
# Load data
data(new_surveys_wide)
}
\author{
Javier Álvarez-Liébana and David Pereiro-Pol.
}
\keyword{datasets}
