% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{next_election_polling_data}
\alias{next_election_polling_data}
\title{Surveys data for congress elections}
\format{
A data frame with 3576 rows and 8 variables:
\itemize{
\item \code{polling_firm}: organisation conducting the poll
\item \code{media}: commissioning organisation / media outlet
\item \code{fieldwork_start}, \code{fieldwork_end},
\code{n_field_days}: fieldwork period and its length (days)
\item \code{sample_size}: sample size of the survey
\item \code{party}: short label of the electoral
\item \code{estimated seats}: estimated number of seats for each party
}
}
\source{
Data extracted from multiple Wikipedia pages
}
\usage{
data(next_election_polling_data)
}
\description{
A dataset containing surveys about the forthcoming congress elections
in Spain.
}
\examples{
# Load data
data(next_election_polling_data)
}
\author{
Javier Álvarez-Liébana and David Pereiro-Pol.
}
\keyword{datasets}
