% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollspaindata-package.R
\docType{package}
\name{pollspaindata-package}
\alias{pollspaindata}
\alias{pollspaindata-package}
\title{pollspaindata: Data companion for the pollspain package}
\description{
The package provides datasets related to electoral results and political surveys in Spain. It is designed to complement the pollspain R package, which offers tools for analyzing Spanish polling data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dadosdelaplace/pollspaindata}
  \item Report bugs at \url{https://github.com/dadosdelaplace/pollspaindata/issues}
}

}
\author{
\strong{Maintainer}: Javier Álvarez-Liébana \email{javalv09@ucm.es} (\href{https://orcid.org/0000-0003-0671-3856}{ORCID}) [contributor, copyright holder]

Authors:
\itemize{
  \item David Pereiro-Pol [contributor]
  \item Mafalda González-González [contributor]
  \item Irene Bosque-Gala [contributor]
}

}
\keyword{internal}
