% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_poll_stations_congress_2019_04}
\alias{raw_poll_stations_congress_2019_04}
\title{Raw data}
\format{
...
}
\source{
Data extracted from
\href{https://infoelectoral.interior.gob.es/opencms/es/elecciones-celebradas/area-de-descargas/}{Spanish Ministry of Interior}
}
\usage{
data(raw_poll_stations_congress_2019_04)
}
\description{
...
}
\examples{
# Load data
data(raw_poll_stations_congress_2019_04)
}
\author{
Javier Álvarez-Liébana and David Pereiro-Pol.
}
\keyword{datasets}
