% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historical_surveys}
\alias{historical_surveys}
\title{Surveys data for congress elections}
\format{
A data frame with 3475 rows and 72 variables:
\itemize{
\item \code{id_elec}: election's id constructed from the election code
and date
\item \code{polling_firm}: organisation conducting the poll
\item \code{media}: commissioning organisation / media outlet
\item \code{fieldwork_start}, \code{fieldwork_end},
\code{n_field_days}: fieldwork period and its length (days)
\item \code{sample_size}: sample size of the survey
\item \code{Lead}: seat difference between the top two parties
\item Party columns
\code{PSOE}, \code{PP}, \code{Sumar}, …, \code{UPyD} (60 in total):
estimated number of seats for each party.
}
}
\source{
pending...
}
\usage{
data(historical_surveys)
}
\description{
A dataset containing surveys about all the  congress elections
held in Spain.
}
\examples{
# Load data
data(historical_surveys)
}
\author{
Javier Álvarez-Liébana and David Pereiro-Pol.
}
\keyword{datasets}
